Move Teach a task from the agent sidepanel to the computer overlay.
Match Grok Bot: the per-bot sidepanel keeps screen preview and routines only.
Teaching starts from a compact control on the live computer view. Web E2E opens
the sidepanel without Teach a task and still covers teaching on the overlay.
