+++ apps/web/e2e/teach-task.spec.ts
@@ -1,11 +1,23 @@
 import { expect, test } from "@playwright/test";
-import { completeOnboarding, signup } from "./helpers";
+import { captureScreenshot, completeOnboarding, signup } from "./helpers";
 
-test("teach a task records interaction and saves a draft", async ({ page }) => {
+test("teach a task records interaction and saves a draft", async ({ page }, testInfo) => {
   const stamp = Date.now();
   await signup(page, `teach-${stamp}@rakazo.test`, "password12", "Teach");
   await completeOnboarding(page);
+
   await page.getByTitle("Agent computer").click();
+  const sidePanel = page.getByTestId("side-panel");
+  await expect(sidePanel).toHaveAttribute("data-panel", "computer");
+  await expect(sidePanel.getByText("Teach a task")).toHaveCount(0);
+  await expect(sidePanel.getByTestId("teach-start-button")).toHaveCount(0);
+  await captureScreenshot(page, testInfo, "teach-sidepanel-no-teach");
+
+  await sidePanel.getByRole("button", { name: "Take control" }).click();
+  await expect(page.getByRole("button", { name: "Close computer" })).toBeVisible();
+  await expect(page.getByTestId("teach-start-button")).toBeVisible();
+  await captureScreenshot(page, testInfo, "teach-computer-overlay");
+
   await page.getByTestId("teach-start-button").click();
   await page.getByTestId("teach-goal-input").fill("Export weekly CRM list");
   await page.getByRole("button", { name: "Start recording" }).click();
